fix: small fixes after a review on video component
